Server/requirements.txt
amqp==2.2.2
appnope==0.1.0
astroid==1.5.3
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.5
djangorestframework==3.6.4
idna==2.6
ipdb==0.10.3
ipython==6.2.0
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.10.2
kombu==4.1.0
lazy-object-proxy==1.3.1
ldap3==2.3
mccabe==0.6.1
osmapi==1.0.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.3.4
Pygments==2.2.0
pyldap==2.4.37
pylint==1.7.2
pyscard==1.9.6
python-can==2.0.0b1
python-dateutil==2.6.1
pytz==2017.2
redis==2.10.6
requests==2.18.4
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.14
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
wrapt==1.10.11
